autotest_rest_api_get_post/requirements.txt
allure-pytest>=2.8.6
PyHamcrest>=1.9.0
pytest>=3.6.1
requests>=2.18.4
selenium>=3.14.0
webdriver-manager>=2.3
xmltodict>=0.12